Add flag to async example (don't compile unless all features are requested)
