+++ examples/simple_radius_server.rs
@@ -1,6 +1,12 @@
-//! Simple example of RADIUS server
+//! An example on how to use RADIUS Server
+//! Contains both Sync & Async RADIUS Server versions
 //!
-//! To try out the example run
+//! To try out Async RADIUS Server example, run
+//! ```bash
+//! cargo run --example simple_radius_server --all-features
+//! ```
+//!
+//! To try out Sync RADIUS Server example, run
 //! ```bash
 //! cargo run --example simple_radius_server
 //! ```
@@ -9,15 +15,21 @@
 use radius_rust::protocol::dictionary::Dictionary;
 use radius_rust::protocol::error::RadiusError;
 use radius_rust::protocol::radius_packet::{ RadiusAttribute, RadiusMsgType, TypeCode };
-use radius_rust::servers::server::{ Server };
 use radius_rust::tools::{ ipv6_string_to_bytes, ipv4_string_to_bytes, integer_to_bytes };
+
+#[cfg(all(feature = "async-radius"))]
+use async_std::task;
+#[cfg(all(feature = "async-radius"))]
+use radius_rust::servers::async_server::AsyncServer as Server;
+#[cfg(all(not(feature = "async-radius")))]
+use radius_rust::servers::server::Server;
 
 
 // Define your own RADIUS packet handlers
-//
-// Ideally, each handler should only return RADIUS packet, that would be send as a response to
-// client
-// In case of error, nothing would be sent to client (which isn't correct behaviour and should be
+// 
+// Ideally, on success, each handler should return RADIUS packet, that would be send as a response to
+// RADIUS client
+// In case of an error, nothing would be sent to client (which isn't correct behaviour and should be
 // fixed later)
 fn handle_auth_request(server: &Server, request: &mut [u8]) -> Result<Vec<u8>, RadiusError> {
     let ipv6_bytes = ipv6_string_to_bytes("fc66::1/64")?;
@@ -47,7 +59,24 @@ fn handle_coa_request(server: &Server, request: &mut [u8]) -> Result<Vec<u8>, Ra
 }
 // ------------------------
 
+#[cfg(all(feature = "async-radius"))]
+fn main() -> Result<(), RadiusError> {
+    task::block_on(async {
+        let dictionary = Dictionary::from_file("./dict_examples/integration_dict")?;
+        let mut server = Server::initialise_server(1812u16, 1813u16, 3799u16, dictionary, String::from("127.0.0.1"), String::from("secret"), 1u16, 2u16).await?;
 
+        server.add_allowed_hosts(String::from("127.0.0.1"));
+
+        server.add_request_handler(RadiusMsgType::AUTH, handle_auth_request)?;
+        server.add_request_handler(RadiusMsgType::ACCT, handle_acct_request)?;
+        server.add_request_handler(RadiusMsgType::COA,  handle_coa_request)?;
+
+        server.run_server().await?;
+        Ok(())
+    })
+}
+
+#[cfg(all(not(feature = "async-radius")))]
 fn main() -> Result<(), RadiusError> {
     let dictionary = Dictionary::from_file("./dict_examples/integration_dict")?;
     let mut server = Server::initialise_server(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2)?;
